Provides group affinity partitioning capability

+++ pkg/controllers/job/job_controller_handler.go
@@ -175,20 +175,28 @@ func (cc *jobcontroller) addPod(obj interface{}) {
 		return
 	}
 
+	taskName, found := pod.Annotations[batch.TaskSpecKey]
+	if !found {
+		klog.Infof("Failed to find taskName of Pod <%s/%s>, skipping",
+			pod.Namespace, pod.Name)
+		return
+	}
+
 	if pod.DeletionTimestamp != nil {
 		cc.deletePod(pod)
 		return
 	}
 
 	req := apis.Request{
-		Namespace: pod.Namespace,
-		JobName:   jobName,
-		JobUid:    jobUid,
-		PodName:   pod.Name,
-		PodUID:    pod.UID,
-
-		Event:      bus.PodPendingEvent,
-		JobVersion: int32(dVersion),
+		Namespace:      pod.Namespace,
+		JobName:        jobName,
+		JobUid:         jobUid,
+		PodName:        pod.Name,
+		PodUID:         pod.UID,
+		TaskName:       taskName,
+		PartitionGroup: getPodPartitionGroup(pod),
+		Event:          bus.PodPendingEvent,
+		JobVersion:     int32(dVersion),
 	}
 
 	if err := cc.cache.AddPod(pod); err != nil {
@@ -298,16 +306,16 @@ func (cc *jobcontroller) updatePod(oldObj, newObj interface{}) {
 	}
 
 	req := apis.Request{
-		Namespace: newPod.Namespace,
-		JobName:   jobName,
-		JobUid:    jobUid,
-		TaskName:  taskName,
-		PodName:   newPod.Name,
-		PodUID:    newPod.UID,
-
-		Event:      event,
-		ExitCode:   exitCode,
-		JobVersion: int32(dVersion),
+		Namespace:      newPod.Namespace,
+		JobName:        jobName,
+		JobUid:         jobUid,
+		TaskName:       taskName,
+		PodName:        newPod.Name,
+		PodUID:         newPod.UID,
+		PartitionGroup: getPodPartitionGroup(newPod),
+		Event:          event,
+		ExitCode:       exitCode,
+		JobVersion:     int32(dVersion),
 	}
 
 	key := jobhelpers.GetJobKeyByReq(&req)
@@ -368,15 +376,15 @@ func (cc *jobcontroller) deletePod(obj interface{}) {
 	}
 
 	req := apis.Request{
-		Namespace: pod.Namespace,
-		JobName:   jobName,
-		JobUid:    jobUid,
-		TaskName:  taskName,
-		PodName:   pod.Name,
-		PodUID:    pod.UID,
-
-		Event:      bus.PodEvictedEvent,
-		JobVersion: int32(dVersion),
+		Namespace:      pod.Namespace,
+		JobName:        jobName,
+		JobUid:         jobUid,
+		TaskName:       taskName,
+		PodName:        pod.Name,
+		PodUID:         pod.UID,
+		PartitionGroup: getPodPartitionGroup(pod),
+		Event:          bus.PodEvictedEvent,
+		JobVersion:     int32(dVersion),
 	}
 
 	if err := cc.cache.DeletePod(pod); err != nil {
